feat: clean bug
- 버그 소멸
