+++ src/main/java/com/timcooki/jnuwiki/domain/docsRequest/dto/request/EditWriteReqDTO.java
@@ -6,7 +6,7 @@ import com.timcooki.jnuwiki.domain.docsRequest.entity.DocsRequestType;
 import lombok.Builder;
 
 public record EditWriteReqDTO(
-        Long id,
+        Long docsId,
         DocsRequestType docsRequestType,
         DocsCategory docsRequestCategory,
         String docsRequestName,
