+++ src/main/java/com/timcooki/jnuwiki/domain/member/controller/AdminController.java
@@ -7,26 +7,17 @@ import com.timcooki.jnuwiki.domain.member.DTO.response.admin.EditListReadResDTO;
 import com.timcooki.jnuwiki.domain.member.DTO.response.admin.EditReadResDTO;
 import com.timcooki.jnuwiki.util.ApiUtils;
 import com.timcooki.jnuwiki.domain.docsRequest.service.DocsRequestService;
-import com.timcooki.jnuwiki.domain.member.DTO.response.admin.EditListReadResDTO;
-import com.timcooki.jnuwiki.domain.member.DTO.response.admin.EditReadResDTO;
-import com.timcooki.jnuwiki.domain.member.DTO.response.admin.NewListReadResDTO;
-import com.timcooki.jnuwiki.domain.member.DTO.response.admin.NewReadResDTO;
 import com.timcooki.jnuwiki.domain.member.service.AdminService;
 import com.timcooki.jnuwiki.domain.member.service.MemberService;
-import com.timcooki.jnuwiki.util.ApiUtils;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Collection;
 
 @RestController
 @RequiredArgsConstructor
@@ -52,11 +43,9 @@ public class AdminController {
     private Object getCreatedRequests(@AuthenticationPrincipal UserDetails userDetails, @PathVariable int page,
                                       @PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
         // 권한 확인
-        Object checkAuthorization = checkAuthorization(userDetails);
-        if (checkAuthorization != null) return checkAuthorization;
 
         // 요청 목록 조회
-        Page<NewListReadResDTO> createRequests = docsRequestService.getCreatedRequestList(pageable);
+        NewListReadResDTO createRequests = docsRequestService.getCreatedRequestList(pageable);
         return ApiUtils.success(createRequests);
     }
 
@@ -75,13 +64,11 @@ public class AdminController {
 
     // 새 문서 신청 요청 상세 조회
     @GetMapping("/requests/new/{docs_request_id}")
-    private Object getOneCreatedRequest(@AuthenticationPrincipal UserDetails userDetails, @PathVariable("docs_request_id") String docsRequestId) {
+    private Object getOneCreatedRequest(@AuthenticationPrincipal UserDetails userDetails, @PathVariable("docs_request_id") Long docsRequestId) {
         // 권한 확인
 
 
         // 요청 존재 확인
-        Object checkRequest = checkValidRequest(docsRequestId);
-        if (checkRequest != null) return checkRequest;
 
         NewReadResDTO modifiedRequest = docsRequestService.getOneCreatedRequest(docsRequestId);
         return ApiUtils.success(modifiedRequest);
