+++ src/main/java/com/timcooki/jnuwiki/domain/member/service/AdminService.java
@@ -1,8 +1,10 @@
 package com.timcooki.jnuwiki.domain.member.service;
 
-import com.timcooki.jnuwiki.domain.docs.DTO.DocsCreateDTO;
+import com.timcooki.jnuwiki.domain.docs.DTO.response.InfoEditResDTO;
+import com.timcooki.jnuwiki.domain.docs.DTO.response.NewApproveResDTO;
 import com.timcooki.jnuwiki.domain.docs.entity.Docs;
 import com.timcooki.jnuwiki.domain.docs.repository.DocsRepository;
+import com.timcooki.jnuwiki.domain.docsArchive.repository.DocsArchiveRepository;
 import com.timcooki.jnuwiki.domain.docsRequest.entity.DocsRequest;
 import com.timcooki.jnuwiki.domain.docsRequest.repository.DocsRequestRepository;
 import com.timcooki.jnuwiki.domain.member.entity.Member;
@@ -11,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,6 +24,7 @@ public class AdminService {
     private final MemberRepository memberRepository;
     private final DocsRequestRepository docsRequestRepository;
     private final DocsRepository docsRepository;
+    private final DocsArchiveRepository docsArchiveRepository;
 
     // TODO - 예외처리 수정
     // 새 문서 요청 승락
@@ -45,21 +49,50 @@ public class AdminService {
                 .createdBy(member)
                 .docsCategory(docsRequest.get().getDocsCategory())
                 .docsLocation(docsRequest.get().getDocsLocation())
-                .docsName(docsRequest.get().getDocsLocation())
+                .docsName(docsRequest.get().getDocsName())
                 .build();
         docsRepository.save(docs);
 
         // DocsRequest 삭제
         docsRequestRepository.deleteById(docsRequestId);
 
-        return ApproveNewDocsResDTO.builder()
+        return NewApproveResDTO.builder()
                 .id(docs.getDocsId())
                 .docsCategory(docs.getDocsCategory())
                 .docsName(docs.getDocsName())
-                .docsLocation(List.of(docs.getDocsLocation()))
+                .docsLocation(docs.getDocsLocation())
                 .build();
 
         //DocsCreateDTO createdDocs = docsRequestService.createDocsFromRequest(docsRequestId);
+    }
+
+    public InfoEditResDTO updateDocsFromRequest(Long docsRequestId) {
+        // 승락받은 요청 조회
+        DocsRequest modifiedRequest = docsRequestRepository.findById(docsRequestId).get();
+
+        // 수정할 문서 조회
+        Long docsId = modifiedRequest.getDocs().getDocsId();
+        Docs docs = docsRepository.findById(docsId).get();
+
+        // 수정전 문서는 아카이브 레포에 저장
+        docsArchiveRepository.save(docs);
+
+        // 요청에 따라 업데이트
+        docs.updateBasicInfo(
+                modifiedRequest.getDocsName(),
+                modifiedRequest.getDocsLocation(),
+                modifiedRequest.getDocsCategory());
+
+        docsRepository.deleteById(docsRequestId); // 처리된 요청 삭제
+
+        return InfoEditResDTO.builder()
+                .docsId(docs.getDocsId())
+                .docsName(docs.getDocsName())
+                .docsLocation(docs.getDocsLocation())
+                .docsContent(docs.getDocsContent())
+                .docsCategory(docs.getDocsCategory())
+                .docsModifiedAt(LocalDateTime.now())
+                .build();
     }
 
 
@@ -71,6 +104,4 @@ public class AdminService {
             return false;
         }
     }
-
-
 }
